Fixing windows warnings
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@588020 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/util/LongSequenceGenerator.cpp
@@ -28,7 +28,7 @@ LongSequenceGenerator::LongSequenceGenerator() {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-int LongSequenceGenerator::getNextSequenceId() {
+long long LongSequenceGenerator::getNextSequenceId() {
     synchronized( &mutex ) {
         return this->lastSequenceId++;
     }
@@ -37,7 +37,7 @@ int LongSequenceGenerator::getNextSequenceId() {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-int LongSequenceGenerator::getLastSequenceId() {
+long long LongSequenceGenerator::getLastSequenceId() {
     synchronized( &mutex ) {
         return this->lastSequenceId;
     }
